+++ src/modifier/query/SQLDistinctModifier.php
@@ -1,0 +1,10 @@
+<?php
+
+trait SQLDistinctModifier
+{
+  public function distinct()
+  {
+    $this->getQuery($this->query())->addModifier('distinct');
+    return $this;
+  }
+}